chore: moved schema to database module
