+++ src/lib.rs
@@ -4,8 +4,8 @@
 extern crate tracing;
 
 pub mod config;
+pub mod database;
 pub mod routes;
-pub mod schema;
 
 pub mod controllers;
 pub mod error;
@@ -13,6 +13,7 @@ pub mod models;
 
 use axum::extract::FromRef;
 pub use config::Config;
+pub use database::schema;
 use deadpool_diesel::postgres::{Object, Pool};
 
 /// An entire database pool
