--- tests/rustc-ice-2025-03-07T17_04_17-1279097.txt
@@ -1,31 +1,0 @@
-thread 'main' panicked at /rust/deps/annotate-snippets-0.9.2/src/display_list/from_snippet.rs:275:9:
-SourceAnnotation range `(80, 94)` is bigger than source length `91`
-stack backtrace:
-   0:     0x7150285cb1e5 - std::backtrace::Backtrace::create::h7208eedda1338435
-   1:     0x7150269a9105 - std::backtrace::Backtrace::force_capture::hee67edab44cd7c83
-   2:     0x715025b199be - std[20d29db3e999c305]::panicking::update_hook::<alloc[e1570dfcc331507b]::boxed::Box<rustc_driver_impl[633be37e059aaf16]::install_ice_hook::{closure#1}>>::{closure#0}
-   3:     0x7150269c1a13 - std::panicking::rust_panic_with_hook::hb92b3172308f4e8b
-   4:     0x7150269c170a - std::panicking::begin_panic_handler::{{closure}}::hbf0e4f1d78818dec
-   5:     0x7150269bef39 - std::sys::backtrace::__rust_end_short_backtrace::hfc58a21f6afae4ae
-   6:     0x7150269c13cd - rust_begin_unwind
-   7:     0x71502365be40 - core::panicking::panic_fmt::h52a3de72b5c008c5
-   8:     0x5a1c0312517d - <annotate_snippets[4274ade940a36283]::display_list::structs::DisplayList as core[7509374bb9bacf81]::convert::From<annotate_snippets[4274ade940a36283]::snippet::Snippet>>::from
-   9:     0x5a1c02f50892 - <rustfmt_nightly[8bef2c5348bbd989]::format_report_formatter::FormatReportFormatter as core[7509374bb9bacf81]::fmt::Display>::fmt
-  10:     0x7150272057e2 - core::fmt::write::h73a1e37ddefe0a4f
-  11:     0x7150269b2dfe - <&std::io::stdio::Stderr as std::io::Write>::write_fmt::h0f2abac5f5e3548b
-  12:     0x7150269b3818 - std::io::stdio::_eprint::hbdd3b0ab288005db
-  13:     0x5a1c02e4f24f - rustfmt[ce1aaa3c9c957e5c]::format_and_emit_report::<std[20d29db3e999c305]::io::stdio::Stdout>
-  14:     0x5a1c02e4dcef - rustfmt[ce1aaa3c9c957e5c]::execute
-  15:     0x5a1c02e48bdf - rustfmt[ce1aaa3c9c957e5c]::main
-  16:     0x5a1c02e3f443 - std[20d29db3e999c305]::sys::backtrace::__rust_begin_short_backtrace::<fn(), ()>
-  17:     0x5a1c02e3f9b9 - std[20d29db3e999c305]::rt::lang_start::<()>::{closure#0}
-  18:     0x71502812699e - std::rt::lang_start_internal::h71fef5e5f25ac3f5
-  19:     0x5a1c02e50278 - main
-  20:     0x71502210a488 - <unknown>
-  21:     0x71502210a54c - __libc_start_main
-  22:     0x5a1c02e30cf9 - <unknown>
-  23:                0x0 - <unknown>
-
-
-rustc version: 1.87.0-nightly (e16a049ad 2025-03-03)
-platform: x86_64-unknown-linux-gnu
